+++ .gitignore
@@ -8,27 +8,30 @@
 *.pvtu
 *.so
 *.vtu
-/firedrake/dmplex.c
-/firedrake/extrusion_numbering.c
-/firedrake/spatialindex.c
-/firedrake/hdf5interface.c
+__pycache__/
+/demos/**/*.py
+/demos/immersed_fem/immersed_domain.msh
 /docs/build/
-/docs/source/firedrake.mg.rst
-/docs/source/firedrake.slope_limiter.rst
-/docs/source/firedrake.matrix_free.rst
-/docs/source/firedrake.preconditioners.rst
+/docs/notebooks/.ipynb_checkpoints/
 /docs/source/firedrake.rst
+/docs/source/firedrake.cython.rst
+/docs/source/firedrake.matrix_free.rst
+/docs/source/firedrake.mg.rst
+/docs/source/firedrake.preconditioners.rst
+/docs/source/firedrake.slate.rst
+/docs/source/firedrake.slate.slac.rst
+/docs/source/firedrake.slate.static_condensation.rst
+/docs/source/firedrake.slope_limiter.rst
 /docs/source/obtaining_pyop2.rst
 /docs/source/teamgrid.rst
 /docs/source/teaminactive.rst
 /docs/source/demos
-/demos/**/*.py
-/firedrake/mg/impl.c
+/firedrake/cython/dmplex.c
+/firedrake/cython/extrusion_numbering.c
+/firedrake/cython/hdf5interface.c
+/firedrake/cython/mgimpl.c
+/firedrake/cython/patchimpl.c
+/firedrake/cython/spatialindex.c
+/firedrake/cython/supermeshimpl.c
 /firedrake_configuration/configuration.json
-/docs/source/firedrake.slate.rst
-/docs/source/firedrake.slate.slac.rst
-/demos/immersed_fem/immersed_domain.msh
-/docs/source/firedrake.slate.static_condensation.rst
-/firedrake/supermeshimpl.c
 /firedrake.egg-info
-__pycache__/
