+++ firedrake/checkpointing.py
@@ -1,6 +1,6 @@
 from firedrake.petsc import PETSc
 from pyop2.mpi import COMM_WORLD, dup_comm, free_comm
-from firedrake import hdf5interface as h5i
+from firedrake.cython import hdf5interface as h5i
 import firedrake
 import numpy as np
 import os
