+++ firedrake/cython/spatialindex.pyx
