+++ firedrake/cython/spatialindexinc.pxi
